fix create user from userlist
